Move shared function to utils.
