+++ lib/extract/effectiveDirective.js
@@ -1,30 +1,17 @@
 'use strict';
 
 var isValidEffectiveDirective = require('../util').isValidEffectiveDirective;
-
-function getEffectiveDirectiveFromEffectiveDirective(payload) {
-  var effectiveDirective = '';
-
-  if (payload.hasOwnProperty('csp-report') && payload['csp-report'].hasOwnProperty('effective-directive') && isValidEffectiveDirective(payload['csp-report']['effective-directive'])) {
-    effectiveDirective = payload['csp-report']['effective-directive'];
-  }
-
-  return effectiveDirective;
-}
-
-function getEffectiveDirectiveFromDocumentURL(payload) {
-  var documentURL = '';
-
-  if (payload.hasOwnProperty('document-url')) {
-      documentURL = payload['document-url'];
-  }
-
-  return documentURL;
-}
+var getEffectiveDirectiveFromEffectiveDirective = require('../util').getEffectiveDirectiveFromEffectiveDirective;
 
 function getEffectiveDirective(payload) {
-  var effectiveDirective = getEffectiveDirectiveFromEffectiveDirective(payload);
-  var documentURL = getEffectiveDirectiveFromDocumentURL(payload);
+  var effectiveDirective = '';
+  var candidateEffectiveDirective = getEffectiveDirectiveFromEffectiveDirective(payload);
+
+  if ('' === candidateEffectiveDirective) {
+
+  } else {
+
+  }
 
   return effectiveDirective;
 }
@@ -45,8 +32,6 @@ function generateEffectiveDirectiveFromViolatedDirective(violatedDirective) {
 }
 
 module.exports = {
-  getEffectiveDirectiveFromEffectiveDirective: getEffectiveDirectiveFromEffectiveDirective,
-  getEffectiveDirectiveFromDocumentURL: getEffectiveDirectiveFromDocumentURL,
   getEffectiveDirective: getEffectiveDirective,
   generateEffectiveDirectiveFromViolatedDirective: generateEffectiveDirectiveFromViolatedDirective
 }
