+++ lib/util.js
@@ -97,10 +97,21 @@ function isValidEffectiveDirective(effectiveDirective) {
   return result;
 }
 
+function getEffectiveDirectiveFromEffectiveDirective(payload) {
+  var effectiveDirective = '';
+
+  if (payload.hasOwnProperty('csp-report') && payload['csp-report'].hasOwnProperty('effective-directive') && isValidEffectiveDirective(payload['csp-report']['effective-directive'])) {
+    effectiveDirective = payload['csp-report']['effective-directive'];
+  }
+
+  return effectiveDirective;
+}
+
 module.exports = {
   getGloballyUniqueIdentifiers: getGloballyUniqueIdentifiers,
   isGloballyUniqueIdentifier: isGloballyUniqueIdentifier,
   sanitizeGloballyUniqueIdentifier: sanitizeGloballyUniqueIdentifier,
   getDirectives: getDirectives,
-  isValidEffectiveDirective: isValidEffectiveDirective
+  isValidEffectiveDirective: isValidEffectiveDirective,
+  getEffectiveDirectiveFromEffectiveDirective: getEffectiveDirectiveFromEffectiveDirective
 };
