+++ test/extract/effectiveDirective.js
@@ -3,8 +3,6 @@
 var assert = require('chai').assert;
 var getEffectiveDirective = require('../../lib/extract/effectiveDirective').getEffectiveDirective;
 var generateEffectiveDirectiveFromViolatedDirective = require('../../lib/extract/effectiveDirective').generateEffectiveDirectiveFromViolatedDirective;
-var getEffectiveDirectiveFromEffectiveDirective = require('../../lib/extract/effectiveDirective').getEffectiveDirectiveFromEffectiveDirective;
-var getEffectiveDirectiveFromDocumentURL = require('../../lib/extract/effectiveDirective').getEffectiveDirectiveFromDocumentURL;
 
 suite(__dirname.split('/').pop(), function() {
   suite(__filename.split('/').pop().replace('.js', ''), function() {
@@ -14,44 +12,6 @@ suite(__dirname.split('/').pop(), function() {
 
     test('generateEffectiveDirectiveFromViolatedDirective is a function', function() {
       assert.isFunction(generateEffectiveDirectiveFromViolatedDirective);
-    });
-
-    test('getEffectiveDirectiveFromEffectiveDirective is a function', function() {
-      assert.isFunction(getEffectiveDirectiveFromEffectiveDirective);
-    });
-
-    test('getEffectiveDirectiveFromDocumentURL is a function', function() {
-      assert.isFunction(getEffectiveDirectiveFromDocumentURL);
-    });
-
-    test('return effective-directive when getting it and one exists', function() {
-      var effectiveDirective = 'img-src';
-      var payload = {
-        'csp-report': {
-          'effective-directive': effectiveDirective
-        }
-      };
-
-      assert.equal(getEffectiveDirective(payload), effectiveDirective);
-    });
-
-    test('return empty string when getting effectiveDirective and one does not exist', function() {
-      var payload = {
-        'csp-report': {}
-      };
-
-      assert.equal(getEffectiveDirective(payload), '');
-      assert.equal(getEffectiveDirective({}), '');
-    });
-
-    test('return empty string when getting effectiveDirective and one does not exist, but is not valid', function() {
-      var payload = {
-        'csp-report': {
-          'effective-directive': 'blah'
-        }
-      };
-
-      assert.equal(getEffectiveDirective(payload), '');
     });
 
     test('return effective-directive when one exists', function() {
